[Add] Add order service + updateDB

+++ GreenSpace_API/GreenSpace.Application/Features/DesignIdeas/Queries/GetDesignIdeaByFillterQuery.cs
@@ -19,8 +19,8 @@ namespace GreenSpace.Application.Features.DesignIdeas.Queries
         public int PageSize { get; set; }
         public string? Category { get; set; }
         public string? Name { get; set; }
-        public float? MinPrice { get; set; }
-        public float? MaxPrice { get; set; }
+        public decimal? MinPrice { get; set; }
+        public decimal? MaxPrice { get; set; }
         public class QueryHandler : IRequestHandler<GetDesignIdeaByFillterQuery, PaginatedList<DesignIdeaViewModel>>
         {
