+++ GreenSpace_API/GreenSpace.Application/Repositories/IProductRepository.cs
@@ -9,6 +9,6 @@ namespace GreenSpace.Application.Repositories
 {
     public interface IProductRepository : IGenericRepository<Product>
     {
-        Task<List<Product>> Search(string? cate, string? name, float? minPrice, float? maxPrice);
+        Task<List<Product>> Search(string? cate, string? name, decimal? minPrice, decimal? maxPrice);
     }
 }
