+++ GreenSpace_API/GreenSpace.Application/ViewModels/UsersWallets/AddAccountBalanceModel.cs
@@ -4,7 +4,7 @@ namespace GreenSpace.Application.ViewModels.UsersWallets;
 
 public class AddAccountBalanceModel
 {
-    public double Amount { get; set; } = 0;
+    public decimal Amount { get; set; } = 0;
     public string TxnRef { get; set; } = string.Empty;
     public string Source { get; set; } = string.Empty;
     public string Type { get; set; } = WalletLogTypeEnum.Deposit.ToString();
